+++ react/components/AccordionFilterGroup.js
@@ -2,12 +2,16 @@ import React from 'react'
 import { useIntl } from 'react-intl'
 import classNames from 'classnames'
 import { useCssHandles, applyModifiers } from 'vtex.css-handles'
+import { useSearchPage } from 'vtex.search-page-context/SearchPageContext'
+import { usePixel } from 'vtex.pixel-manager'
 
 import AccordionFilterItem from './AccordionFilterItem'
 import FacetCheckboxList from './FacetCheckboxList'
 import useSelectedFilters from '../hooks/useSelectedFilters'
 import { getFilterTitle } from '../constants/SearchHelpers'
 import { searchSlugify } from '../utils/slug'
+import RadioFilters from './RadioFilters'
+import { pushFilterManipulationPixelEvent } from '../utils/filterManipulationPixelEvents'
 
 const CSS_HANDLES = ['accordionFilterOpen']
 
@@ -29,6 +33,7 @@ const AccordionFilterGroup = ({
   onClearFilter,
   showClearByFilter,
 }) => {
+  const { searchQuery } = useSearchPage()
   const handles = useCssHandles(CSS_HANDLES)
   const filters = useSelectedFilters(facets)
   const selectedFilters = filters.filter(facet => facet.selected)
@@ -36,6 +41,9 @@ const AccordionFilterGroup = ({
   const facetTitle = getFilterTitle(title, intl)
   const slugifiedFacetTitle = searchSlugify(facetTitle)
   const facetKey = filters.length > 0 ? filters[0].key : null
+  const { push } = usePixel()
+
+  const isRadio = filters.some(filter => filter.key === 'shipping')
 
   return (
     <AccordionFilterItem
@@ -58,16 +66,31 @@ const AccordionFilterGroup = ({
           className
         )}
       >
-        <FacetCheckboxList
-          onFilterCheck={onFilterCheck}
-          facets={filters}
-          quantity={quantity}
-          facetTitle={facetTitle}
-          truncateFilters={truncateFilters}
-          navigationType={navigationType}
-          truncatedFacetsFetched={truncatedFacetsFetched}
-          setTruncatedFacetsFetched={setTruncatedFacetsFetched}
-        />
+        {isRadio ? (
+          <RadioFilters
+            facets={filters}
+            onChange={facet => {
+              pushFilterManipulationPixelEvent({
+                name: facetTitle,
+                value: facet.name,
+                products: searchQuery?.products ?? [],
+                push,
+              })
+              onFilterCheck({ ...facet, title: facetTitle }, true)
+            }}
+          />
+        ) : (
+          <FacetCheckboxList
+            onFilterCheck={onFilterCheck}
+            facets={filters}
+            quantity={quantity}
+            facetTitle={facetTitle}
+            truncateFilters={truncateFilters}
+            navigationType={navigationType}
+            truncatedFacetsFetched={truncatedFacetsFetched}
+            setTruncatedFacetsFetched={setTruncatedFacetsFetched}
+          />
+        )}
       </div>
     </AccordionFilterItem>
   )